fix: avoid duplicate table names in tests
